Add utils file and exports
