+++ src/index.js
@@ -1,4 +1,10 @@
 // src/index.js
 import SessionManagement from './lib/session-management.js';
+import {
+  createDefaultExpireTimes, checkSessionStatus, renewSession, isSessionExpired, convertUTCToJSDate,
+} from './utils/utils.js';
 
+export {
+  createDefaultExpireTimes, checkSessionStatus, renewSession, isSessionExpired, convertUTCToJSDate,
+};
 export default SessionManagement;
