+++ src/lib/session-management.js
@@ -1,12 +1,15 @@
 import fp from 'lodash/fp.js';
-import { apiConfig, defaultConfig } from '../config/config.js';
+import { defaultConfig } from '../config/config.js';
+import {
+  checkSessionStatus, renewSession, convertUTCToJSDate,
+} from '../utils/utils.js';
 
 class SessionManagement {
   static instance;
 
   constructor() {
     if (SessionManagement.instance) {
-      return SessionManagement.instance;
+      throw new Error('Use SessionManagement.getInstance() to get the single instance of this class.');
     }
 
     this.config = {};
@@ -21,6 +24,13 @@ class SessionManagement {
     SessionManagement.instance = this;
   }
 
+  static getInstance() {
+    if (!SessionManagement.instance) {
+      SessionManagement.instance = new SessionManagement();
+    }
+    return SessionManagement.instance;
+  }
+
   async init(config) {
     if (!config || typeof config !== 'object') {
       throw new Error('[LIBRARY] Invalid configuration object');
@@ -29,11 +39,11 @@ class SessionManagement {
     console.log('[LIBRARY] Initialising session management with config:', this.config);
     if (this.config.checkSessionOnInit) {
       console.log('[LIBRARY] Checking initial session state');
-      const data = await this.checkSessionStatus();
+      const data = await checkSessionStatus();
       const sessionExpiryTime = fp.get('expirationTime')(data);
       if (sessionExpiryTime) {
         console.log('[LIBRARY] Initial session is active, setting timers');
-        this.setSessionExpiryTime(this.convertUTCToJSDate(sessionExpiryTime));
+        this.setSessionExpiryTime(convertUTCToJSDate(sessionExpiryTime));
         if (this.config.onSessionValid) {
           this.config.onSessionValid(sessionExpiryTime);
         }
@@ -85,15 +95,6 @@ class SessionManagement {
     );
   }
 
-  convertUTCToJSDate(expiryTime) {
-    if (expiryTime) {
-      const expireTimeInUTCString = expiryTime.replace(' +0000 UTC', 'Z');
-      console.log('[LIBRARY] UTC to JS date: ', expireTimeInUTCString);
-      return new Date(expireTimeInUTCString);
-    }
-    return null;
-  }
-
   startExpiryTimer(name, expiryTime, offsetInMilliseconds, callback) {
     console.log(`[LIBRARY] Expiry time for ${name}: ${expiryTime}`);
     if (expiryTime) {
@@ -140,9 +141,9 @@ class SessionManagement {
     };
     console.log('[LIBRARY] Updating session timer via API 1');
     try {
-      const response = await this.renewSession();
+      const response = await renewSession();
       if (response) {
-        const expirationTime = this.convertUTCToJSDate(fp.get('expirationTime')(response));
+        const expirationTime = convertUTCToJSDate(fp.get('expirationTime')(response));
         console.log(
           '[LIBRARY] Session renewed successfully, new expiration time:',
           expirationTime,
@@ -159,65 +160,6 @@ class SessionManagement {
     }
   }
 
-  isSessionExpired(sessionExpiryTime) {
-    if (sessionExpiryTime == null) {
-      return true;
-    }
-    console.log('[IS SESSION EXPIRED] sessionExpiryTime: ', sessionExpiryTime);
-    const now = new Date();
-    const nowUTCInMS = now.getTime() + now.getTimezoneOffset() * 60000;
-    const nowInUTC = new Date(nowUTCInMS);
-    console.log('[IS SESSION EXPIRED] nowInUTC: ', nowInUTC);
-    console.log('[IS SESSION EXPIRED] sessionExpiryTime: ', new Date(sessionExpiryTime));
-    // Get the time difference between now and the expiry time minus the timer offset
-    const timerInterval = new Date(sessionExpiryTime) - nowInUTC;
-    console.log('[IS SESSION EXPIRED] timerInterval: ', timerInterval);
-    const diffInSeconds = Math.round(timerInterval / 1000);
-    console.log('[IS SESSION EXPIRED] diff: ', diffInSeconds);
-    if (Number.isNaN(diffInSeconds)) {
-      throw new Error('encounted an error checking time interval: diffInSeconds is NaN');
-    }
-    if (diffInSeconds <= 0) {
-      return true;
-    }
-    return false;
-  }
-
-  async renewSession(body) {
-    console.log('[LIBRARY] Starting session renewal process');
-    const response = await fetch(apiConfig.RENEW_SESSION, {
-      method: 'PUT',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(body),
-    });
-
-    console.log('[LIBRARY] Fetch request sent to /tokens/self: ', response);
-
-    if (!response.ok) {
-      console.error('[LIBRARY] Failed to renew session, response status:', response.status);
-      throw new Error('Failed to renew session');
-    }
-
-    const data = await response.json();
-    console.log('[LIBRARY] Session renewed successfully, response data:', data);
-
-    return data;
-  }
-
-  async checkSessionStatus() {
-    console.log('[LIBRARY] Checking initial session status');
-    const response = await fetch(apiConfig.CHECK_SESSION, { method: 'GET' });
-
-    if (response.ok) {
-      const data = await response.json();
-      console.log('[LIBRARY] Initial session status:', data);
-      return data;
-    }
-    throw new Error('Failed to check session status');
-  };
-
   removeTimers() {
     this.removeInteractionMonitoring();
 
@@ -227,18 +169,8 @@ class SessionManagement {
 
     this.timers = {};
   }
-
-  createDefaultExpireTimes(hours) {
-    console.log('[LIBRARY] Creating default expire times for', hours, 'hours');
-    const now = new Date();
-    const expiry = now.setHours(now.getHours() + hours);
-    return {
-      session_expiry_time: new Date(expiry),
-      refresh_expiry_time: new Date(expiry),
-    };
-  }
 }
 
 // Export a single instance of SessionManagement
-const sessionManagementInstance = new SessionManagement();
+const sessionManagementInstance = SessionManagement.getInstance();
 export default sessionManagementInstance;
